feat(Finalização)

+++ springsecurity/src/main/java/tech/builrun/springsecurity/controllers/dto/FeedDto.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public record FeedDto(List<FeedItemDto> feedItemDtos, int page, int pageSize,
                       int totalPages,
-                      int totalElements) {
+                      Long totalElements) {
 
 }
